refactor: order(common,delivery,eatin,takeout)
