--- src/main/java/kitchenpos/order/application/facade/OrderFacade.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.application.facade;
-
-public class OrderFacade {
-
-}
