+++ src/main/java/kitchenpos/order/common/application/facade/OrderTableFacade.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.common.application.facade;
+
+public class OrderTableFacade {
+
+}
