+++ src/main/java/kitchenpos/order/common/domain/entity/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.common.domain.entity;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
