+++ src/main/java/kitchenpos/order/common/domain/entity/OrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.common.domain.entity;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
