+++ src/main/java/kitchenpos/order/common/domain/repository/OrderTableRepository.java
@@ -1,0 +1,16 @@
+package kitchenpos.order.common.domain.repository;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+import kitchenpos.order.common.domain.entity.OrderTable;
+
+public interface OrderTableRepository {
+
+    OrderTable save(OrderTable orderTable);
+
+    Optional<OrderTable> findById(UUID id);
+
+    List<OrderTable> findAll();
+}
+
