+++ src/main/java/kitchenpos/order/common/infrastructure/persistence/JpaOrderRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.order.common.infrastructure.persistence;
+
+import java.util.UUID;
+import kitchenpos.order.common.domain.entity.Order;
+import kitchenpos.order.common.domain.repository.OrderRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
+
+}
