+++ src/main/java/kitchenpos/order/common/infrastructure/persistence/JpaOrderTableRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.order.common.infrastructure.persistence;
+
+import java.util.UUID;
+import kitchenpos.order.common.domain.entity.OrderTable;
+import kitchenpos.order.common.domain.repository.OrderTableRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaOrderTableRepository extends OrderTableRepository,
+    JpaRepository<OrderTable, UUID> {
+
+}
