+++ src/main/java/kitchenpos/order/delivery/domain/dto/DeliveryDto.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.delivery.domain.dto;
+
+public record DeliveryDto() {
+
+}
