+++ src/main/java/kitchenpos/order/delivery/domain/service/DeliveryKitchenridersClient.java
@@ -1,0 +1,9 @@
+package kitchenpos.order.delivery.domain.service;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface DeliveryKitchenridersClient {
+
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
