--- src/main/java/kitchenpos/order/domain/entity/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.domain.entity;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
-}
