--- src/main/java/kitchenpos/order/domain/entity/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.domain.entity;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
