--- src/main/java/kitchenpos/order/domain/repository/OrderTableRepository.java
@@ -1,16 +1,0 @@
-package kitchenpos.order.domain.repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import kitchenpos.order.domain.entity.OrderTable;
-
-public interface OrderTableRepository {
-
-    OrderTable save(OrderTable orderTable);
-
-    Optional<OrderTable> findById(UUID id);
-
-    List<OrderTable> findAll();
-}
-
