--- src/main/java/kitchenpos/order/domain/service/DeliveryKitchenridersClient.java
@@ -1,9 +1,0 @@
-package kitchenpos.order.domain.service;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public interface DeliveryKitchenridersClient {
-
-    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
-}
