+++ src/main/java/kitchenpos/order/eatin/domain/dto/EatinDto.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.eatin.domain.dto;
+
+public record EatinDto() {
+
+}
