+++ src/main/java/kitchenpos/order/takeout/domain/dto/TakeoutDto.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.takeout.domain.dto;
+
+public record TakeoutDto() {
+
+}
