+++ src/main/java/kitchenpos/order/takeout/domain/service/TakeoutService.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.takeout.domain.service;
+
+public class TakeoutService {
+
+}
