+++ src/test/java/kitchenpos/order/application/OrderFacadeTest.java
@@ -17,18 +17,18 @@ import java.util.stream.Collectors;
 import kitchenpos.menu.domain.entity.Menu;
 import kitchenpos.menu.domain.fixture.MenuFixture;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.order.domain.entity.Order;
-import kitchenpos.order.domain.entity.OrderLineItem;
-import kitchenpos.order.domain.entity.OrderStatus;
-import kitchenpos.order.domain.entity.OrderTable;
-import kitchenpos.order.domain.entity.OrderType;
+import kitchenpos.order.common.domain.entity.Order;
+import kitchenpos.order.common.domain.entity.OrderLineItem;
+import kitchenpos.order.common.domain.entity.OrderStatus;
+import kitchenpos.order.common.domain.entity.OrderTable;
+import kitchenpos.order.common.domain.entity.OrderType;
 import kitchenpos.order.domain.fixture.OrderFixture;
 import kitchenpos.order.domain.fixture.OrderLineItemFixture;
 import kitchenpos.order.domain.fixture.OrderTableFixture;
-import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order.domain.repository.OrderTableRepository;
-import kitchenpos.order.domain.service.DeliveryKitchenridersClient;
-import kitchenpos.order.domain.service.OrderService;
+import kitchenpos.order.common.domain.repository.OrderRepository;
+import kitchenpos.order.common.domain.repository.OrderTableRepository;
+import kitchenpos.order.delivery.domain.service.DeliveryKitchenridersClient;
+import kitchenpos.order.common.domain.service.OrderService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
