+++ src/test/java/kitchenpos/order/application/OrderTableFacadeTest.java
@@ -9,12 +9,12 @@ import static org.mockito.Mockito.when;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.order.domain.entity.OrderStatus;
-import kitchenpos.order.domain.entity.OrderTable;
+import kitchenpos.order.common.domain.entity.OrderStatus;
+import kitchenpos.order.common.domain.entity.OrderTable;
 import kitchenpos.order.domain.fixture.OrderTableFixture;
-import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order.domain.repository.OrderTableRepository;
-import kitchenpos.order.domain.service.OrderTableService;
+import kitchenpos.order.common.domain.repository.OrderRepository;
+import kitchenpos.order.common.domain.repository.OrderTableRepository;
+import kitchenpos.order.eatin.domain.service.OrderTableService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
